+++ packages/redwoodjs/src/ratelimit/middleware/__tests__/util.test.ts
@@ -1,6 +1,6 @@
 import { MiddlewareRequest } from "@redwoodjs/vite/middleware";
 import { assert, describe, expect, it } from "vitest";
-import { defaultRatelimitIdentifier } from "../../middleware/util";
+import { defaultRatelimitIdentifier } from "../util";
 
 describe("defaultRatelimitIdentifier", () => {
   it("should return correct identifier", () => {
